Introduced ValidatesFileTrait for file validation, ensuring security and type restrictions.
